requirements.txt
# Web Framework
Flask>=3.0.0
streamlit>=1.28.0

# AI/ML
google-generativeai>=0.6.0

# Image Processing
Pillow>=10.0.0

# HTTP Requests & Environment
requests>=2.31.0
python-dotenv>=1.0.0

# MCP Server
fastmcp
beautifulsoup4>=4.12.0
lxml>=4.9.0
mcp[cli]>=1.6.0
